清理index类算子编译过程中的告警，修复gather_elements_v2的超大shape场景运行报错与修复算子ophost ut未运行



# message auto-generated for no-merge-commit merge:
!121 merge master into master

清理index类算子编译过程中的告警，修复gather_elements_v2的超大shape场景运行报错与修复算子ophost ut未运行

Created-by: sakuraqqz
Commit-by: 付饶
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
1. 清理index类算子编译过程中的告警
2. 修改gather_elements_v2超大shape下的算子kernel报错，模型场景已验证。
3. 修复算子repeat_interleave_grad，inplace_index_add_with_sorted，inplace_scatter_add的ophost_ut，运行结果已贴评论区。
## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->
https://gitcode.com/cann/ops-nn/issues/56
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!121

+++ index/gather_elements/op_host/op_api/gather_elements.cpp
@@ -114,7 +114,7 @@ static int64_t GetTensorSize(const aclTensor *input) {
 
 static bool IsSameDimValueExceptAxis(const op::Shape x_shape, const op::Shape index_shape, const int64_t axis, const size_t dims) {
   for (size_t i = 0; i < dims; i++) {
-    if ((i != axis) && (x_shape[i] != index_shape[i])) {
+    if (static_cast<int64_t>(i) != axis && (x_shape[i] != index_shape[i])) {
       return false;
     }
   }
@@ -133,7 +133,7 @@ static bool IsLastAxisSupport(const aclTensor *self, const aclTensor *index, con
   int64_t x_axis = x_shape[axis];
   int64_t repeat_per_core = LEAST_REPEAT_TIME;
   bool if_same_dim_value_except_axis = IsSameDimValueExceptAxis(x_shape, index_shape, axis, dims);
-  bool is_last_axis = (axis == dims - 1);
+  bool is_last_axis = (axis == static_cast<int64_t>(dims - 1));
   auto socVersion = GetCurrentPlatformInfo().GetSocVersion();
   bool isSupportSoc = (socVersion == SocVersion::ASCEND910B || socVersion == SocVersion::ASCEND910_93);
   int64_t ubSize = isSupportSoc ? SMALL_UB_SIZE : UB_SIZE;
