+++ index/gather_v2/op_host/op_api/aclnn_gather_v3.cpp
@@ -133,7 +133,7 @@ static aclnnStatus CheckParams(const aclTensor *self, int64_t dim, const aclTens
 
   auto xsize = ge::GetSizeByDataType(self->GetDataType());
   auto selfDimNum = self->GetViewShape().GetDimNum();
-  for (int i = 0; i < selfDimNum ; i++) {
+  for (uint64_t i = 0; i < selfDimNum ; i++) {
     xsize *= self->GetViewShape().GetDim(i);
   }
